cln(utils): split utils into plotting and data
